--- src/main/kotlin/domain/model/LottoDrawingResult.kt
@@ -1,3 +1,0 @@
-package domain.model
-
-data class LottoDrawingResult(val statistics: Map<Rank, Int>)
