--- src/main/kotlin/domain/model/Margin.kt
@@ -1,3 +1,0 @@
-package domain.model
-
-data class Margin(val rate: Double)
