+++ src/main/kotlin/lotto/domain/LottoGenerator.kt
@@ -1,0 +1,18 @@
+package lotto.domain
+
+import lotto.domain.model.Lotto
+import lotto.domain.model.LottoNumber
+
+object LottoGenerator {
+    private const val MINIMUM_LOTTO_NUMBER = 1
+    private const val MAXIMUM_LOTTO_NUMBER = 45
+    private const val LOTTO_SIZE = 6
+
+    fun makeLotto(randomNumber: List<Int>): Lotto {
+        val randomNumbers = randomNumber.take(LOTTO_SIZE).sorted()
+        val lottoNumbers = randomNumbers.map { LottoNumber(it) }
+        return Lotto(lottoNumbers)
+    }
+
+    fun makeRandomNumber() = (MINIMUM_LOTTO_NUMBER..MAXIMUM_LOTTO_NUMBER).shuffled()
+}
