+++ src/main/kotlin/lotto/domain/model/LottoDrawingResult.kt
@@ -1,0 +1,3 @@
+package lotto.domain.model
+
+data class LottoDrawingResult(val statistics: Map<Rank, Int>)
