+++ src/main/kotlin/lotto/domain/model/Margin.kt
@@ -1,0 +1,3 @@
+package lotto.domain.model
+
+data class Margin(val rate: Double)
