+++ src/main/kotlin/lotto/util/Constants.kt
@@ -1,0 +1,5 @@
+package lotto.util
+
+object Constants {
+    const val LOTTO_PRICE = 1000
+}
