--- src/main/kotlin/util/Constants.kt
@@ -1,5 +1,0 @@
-package util
-
-object Constants {
-    const val LOTTO_PRICE = 1000
-}
