--- src/test/kotlin/domain/CashierTest.kt
@@ -1,13 +1,0 @@
-package domain
-
-import domain.model.Money
-import org.assertj.core.api.Assertions.assertThat
-import org.junit.jupiter.api.Test
-
-class CashierTest {
-    @Test
-    fun name() {
-        val result = Cashier().toTicketQuantity(Money(10000))
-        assertThat(result).isEqualTo(10)
-    }
-}
